moving reddit actions into subdirectories
