+++ rules.json
@@ -1,1 +1,1 @@
-
+{"rules": [{"set": {"text": ["Queen Sandra", "President Sarah Lacina", "Troyzan's Island", "F*** you, Brad Culpepper!", "Depth Charge", "Dogs Samsung Chinese lol", "This is a dictatorship - majority rules.", "\"Hero\"", "\"Healer\"", "\"Hustler\""], "limit": {"available": 0, "expires": 9999999999, "usergroup": "all"}, "class": [{"content": "X", "limit": {"available": 0, "expires": 9999999999, "usergroup": "all"}, "type": "string"}, {"content": " W", "type": "string"}, {"content": "(?<=W)(.*?)(?=W)", "type": "regex"}, {"content": "W ", "type": "string"}, {"content": "W.*W(.*)", "type": "regex"}], "setname": "GreyFlairs"}}, {"set": {"text": ["A.K. (AUS)", "Adam (AUS)", "Aimee (AUS)", "Anneliese (AUS)", "Ben (AUS)", "Henry (AUS)", "Jacqui (AUS)", "Jarrad (AUS)", "Jericho (AUS)", "Joan (AUS)", "Kate (AUS)", "Kent (AUS)", "Locky (AUS)", "Luke (AUS)", "Mark H (AUS)", "Tarzan (AUS)", "Mark W (AUS)", "Michelle (AUS)", "Odette (AUS)", "Peter (AUS)", "Sam (AUS)", "Sarah (AUS)", "Tara (AUS)", "Tessa (AUS)", "Ziggy (AUS)"], "limit": {"available": 0, "expires": 9999999999, "usergroup": "all"}, "class": [{"content": "AUS17", "limit": {"available": 0, "expires": 9999999999, "usergroup": "all"}, "type": "string"}, {"content": " W", "type": "string"}, {"content": "(?<=W)(.*?)(?=W)", "type": "regex"}, {"content": "W ", "type": "string"}, {"content": "W.*W(.*)", "type": "regex"}], "setname": "AustralianSurvivor2017"}}, {"set": {"text": ["Alan", "Ben", "Cole", "Devon", "Joe", "John", "Mike", "Patrick", "Ryan", "Alexandrea", "Ashley", "Christine", "Desiree", "Jessica", "Katrina", "Lauren", "Roark", "Simone"], "limit": {"available": 0, "expires": 9999999999, "usergroup": "all"}, "class": [{"content": "35Gold", "limit": {"available": 0, "expires": 1507161600, "usergroup": "all"}, "type": "string"}, {"content": "35Silver", "limit": {"available": 1507161601, "expires": 9999999999, "usergroup": "all"}, "type": "string"}, {"content": " W", "type": "string"}, {"content": "(?<=W)(.*?)(?=W)", "type": "regex"}, {"content": "W ", "type": "string"}, {"content": "W.*W(.*)", "type": "regex"}], "setname": "Season35"}}]}
